Endpoints y otros
Endpoints provisionales y cambio dato puntuaciones a double

+++ database/migrations/2021_04_22_171331_create_puntuacion_productos_table.php
@@ -19,7 +19,7 @@ class CreatePuntuacionProductosTable extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('producto_id')->constrained('productos');
-            $table->integer('puntuacion_producto');
+            $table->double('puntuacion_producto');
             $table->text('comentario');
             $table->timestamps();
         });
